refactor(offchain): file changed and used
refactor(offchain): unnecessary arrow function removed
